+++ Domain/Entities/Shop.cs
@@ -4,17 +4,17 @@ namespace Domain.Entities;
 
 public class Shop
 {
-    public uint Id { get; set; }
+    public long Id { get; set; }
     public Guid Uuid { get; set; }
     public string? IgId { get; set; }
     public required string Name { get; set; }
     public string? Description { get; set; }
     public string? Logo { get; set; }
-    public ushort TotalProducts { get; set; }
+    public int TotalProducts { get; set; }
     public string? IgUsername { get; set; }
     public string? IgFullName { get; set; }
-    public uint? IgFollowers { get; set; }
-    public uint? OwnerId { get; set; }
+    public long? IgFollowers { get; set; }
+    public long? OwnerId { get; set; }
     public bool IsVerified { get; set; }
     public bool IsValidated { get; set; }
     public bool IsDeleted { get; set; }
