+++ Infrastructure/Dtos/ProductDto.cs
@@ -6,7 +6,7 @@ public class ProductDto
 {
     public class ProductResponseDto
     {
-        public ulong Id { get; set; }
+        public long Id { get; set; }
         public Guid Uuid { get; set; }
         public string? IgId { get; set; }
         public required string Name { get; set; }
@@ -14,15 +14,15 @@ public class ProductDto
         public decimal Price { get; set; }
         public decimal OriginalPrice { get; set; }
         public byte DiscountPercentage { get; set; }
-        public uint Stock { get; set; }
+        public long Stock { get; set; }
         public string? IgCode { get; set; }
         public string? IgThumbnailSrc { get; set; }
         public string? IgDisplayUrl { get; set; }
-        public uint? IgLikeCount { get; set; }
-        public uint? IgCommentCount { get; set; }
+        public long? IgLikeCount { get; set; }
+        public long? IgCommentCount { get; set; }
         public byte? IgCarouselMediaCount { get; set; }
         public string? IgVideoUrl { get; set; }
-        public uint ShopId { get; set; }
+        public long ShopId { get; set; }
         public bool HasDiscount { get; set; }
         public bool IsAvailable { get; set; }
         public bool IsNew { get; set; }
@@ -46,27 +46,27 @@ public class ProductDto
         public string? Description { get; set; }
         public decimal OriginalPrice { get; set; }
         public byte DiscountPercentage { get; set; }
-        public uint Stock { get; set; }
-        public uint ShopId { get; set; }
+        public long Stock { get; set; }
+        public long ShopId { get; set; }
         public required List<string> Images { get; set; }
-        public required List<ushort> CategoryIds { get; set; }
+        public required List<int> CategoryIds { get; set; }
     }
 
     public class ProductCreateIgDto
     {
         public List<string> Usernames { get; set; }
-        public List<ushort> CategoryIds { get; set; }
+        public List<int> CategoryIds { get; set; }
     }
 
     public class ProductUpdateDto
     {
         // TODO
-        public ulong Id { get; set; }
+        public long Id { get; set; }
         public string? Name { get; set; }
         public string? Description { get; set; }
         public decimal? Price { get; set; }
         public byte? DiscountPercentage { get; set; }
-        public uint? Stock { get; set; }
+        public long? Stock { get; set; }
         public List<string>? Images { get; set; }
     }
 }
