+++ Infrastructure/Dtos/SharedDto.cs
@@ -4,9 +4,9 @@ public class SharedDto
 {
     public class GetAllDto
     {
-        public ushort PageNumber { get; set; }
-        public ushort PageSize { get; set; }
+        public int PageNumber { get; set; }
+        public int PageSize { get; set; }
         public required string SearchTerm { get; set; }
-        public List<ushort>? CategoryIds { get; set; }
+        public List<int>? CategoryIds { get; set; }
     }
 }
